Final spam commit after publishing this repo
this repo was only on my machine, and when I tried to publish it, github will always disconnect at the end (it just say remote disconnected unexpectedly).

some searching later and this error seemed to happen when sending large files around 2GB (mine was just 447MB).
Anyway, I solved it by sending my files as much as i could each commit (and of-course losing the old history, but at least the code is online now)